feat: move trace related runtime function into TraceRuntime trait.
Note that the types we used in `batch_message_channel` function and `TrySend` trait are all from `trace` mod. But the runtime doesn't live in trace mod. So we use another trait `TraceRuntime` to include all function and types that related `trace` mod
