+++ opentelemetry/src/sdk/trace/mod.rs
@@ -11,6 +11,7 @@ mod evicted_hash_map;
 mod evicted_queue;
 mod id_generator;
 mod provider;
+mod runtime;
 mod sampler;
 mod span;
 mod span_limit;
@@ -22,6 +23,7 @@ pub use evicted_hash_map::EvictedHashMap;
 pub use evicted_queue::EvictedQueue;
 pub use id_generator::{aws::XrayIdGenerator, IdGenerator};
 pub use provider::{Builder, TracerProvider};
+pub use runtime::{TraceRuntime, TrySend};
 pub use sampler::{Sampler, SamplingDecision, SamplingResult, ShouldSample};
 pub use span::Span;
 pub use span_limit::SpanLimits;
